Moved all command related methods to CommandHandler from bot
