+++ src/main/java/se/nikoci/lgdiscordbot/lib/Bot.java
@@ -42,64 +42,4 @@ public class Bot {
         this.commandHandler = commandHandler;
         jda.addEventListener(commandHandler);
     }
-
-    public void addCommand(@NotNull Command command){
-        //restrict adding multiple with the same name
-        for (var entrySet : commands.entrySet()){
-            if (entrySet.getKey().equalsIgnoreCase(command.getName())){
-                logger.error("Cannot add command with the same name");
-                return;
-            }
-        }
-
-        commands.put(command.getName(), command);
-        updateCommandData();
-    }
-
-    public void removeCommand(@NotNull Command command){
-        removeCommand(command.getName());
-    }
-
-    public void removeCommand(@NotNull String commandName){
-        commands.forEach((name, cmd) -> {
-            if (name.equalsIgnoreCase(commandName)) commands.remove(name, cmd);
-        });
-    }
-
-    public Command getCommand(@NotNull String name){
-        for (var entrySet : commands.entrySet()){
-            if (entrySet.getKey().equalsIgnoreCase(name)){
-                return entrySet.getValue();
-            }
-        }
-        return null;
-    }
-
-    public void addCommands(Command @NotNull ... commands){
-        for (Command command : commands) addCommand(command);
-    }
-
-    public void removeCommands(Command @NotNull ... commands){
-        for (Command command : commands) removeCommand(command);
-    }
-
-    //Updates command data for jda, only Command obj's that accept slash commands will be updated.
-    private void updateCommandData(){
-        List<CommandData> commandDataList = new ArrayList<>();
-
-        commands.forEach((name, cmd) -> {
-            //Make sure we only select Command objects that accept slash commands
-            if (cmd.getCommandType().contains(CommandType.GUILD_SLASH)
-                    || cmd.getCommandType().contains(CommandType.DM_SLASH)
-                    || cmd.getCommandType().contains(CommandType.ALL)){
-                commandDataList.add(cmd.getCommandData());
-            }
-        });
-
-        getJda()
-                .updateCommands()
-                .addCommands(commandDataList)
-                .queue();
-
-    }
 }
